+++ src/session.rs
@@ -163,6 +163,26 @@ where
         })
     }
 
+    /// Removes a Key from the Current Session's HashMap returning it.
+    ///
+    /// Provides an Option<T> that returns the requested data from the Sessions store.
+    /// Returns None if Key does not exist or if serdes_json failed to deserialize.
+    ///
+    /// # Examples
+    /// ```rust ignore
+    /// let id = session.get_remove("user-id").await.unwrap_or(0);
+    /// ```
+    ///
+    /// Used to get data stored within SessionDatas hashmap from a key value.
+    ///
+    #[inline]
+    pub async fn get_remove<T: serde::de::DeserializeOwned>(&self, key: &str) -> Option<T> {
+        self.tap(|sess| {
+            let string = sess.data.remove(key)?;
+            serde_json::from_str(&string).ok()
+        })
+    }
+
     /// Sets data to the Current Session's HashMap.
     ///
     /// # Examples
@@ -183,7 +203,8 @@ where
         });
     }
 
-    /// Removes a Key from the Current Session's HashMap returning it.
+    /// Removes a Key from the Current Session's HashMap.
+    /// Does not process the String into a Type, Just removes it.
     ///
     /// # Examples
     /// ```rust ignore
@@ -191,12 +212,11 @@ where
     /// ```
     ///
     #[inline]
-    pub async fn remove<T: serde::de::DeserializeOwned>(&self, key: &str) -> Option<T> {
+    pub async fn remove(&self, key: &str) {
         self.tap(|sess| {
             sess.update = true;
-            let string = sess.data.remove(key)?;
-            serde_json::from_str(&string).ok()
-        })
+            sess.data.remove(key)
+        });
     }
 
     /// Clears all data from the Current Session's HashMap.